Extract subview for panning and zooming logic
